# Improve `vec_equal_na()` performance